fix: hash computation error (critical).
Two function with same logic:

- examples/example_simple_french.py
- examples/example_simple.py

had different hash. When normalizing the functions, variable names
were sorted according to lexicographic order, that lead to different
normal representation of the same logic. In one case:

  def mobius_v_0(mobius_v_2, mobius_v_1):...

In another, the expected:

  def mobius_v_0(mobius_v_1, mobius_v_0):...

Except the function that is always called mobius_v_0, variables must
be renamed from left to right, top to bottom, in order of appearance
in the code, to guarantee same normal form, for same logic, hence
consistent hashing.